# [Opt] speed up vectorized aggregation function