**使用__restrict关键字加速聚合函数计算**

加速并没有原pr里面说的那么显著：https://github.com/ClickHouse/ClickHouse/pull/19946


- 测试集群配置：1fe，3be

- 数据源：TPCH，单表 orders，3000w行数据

- 将不开vec、开启vec优化前、开启vec优化后三组进行对比

poc测试结论：

1. 综合查询下，对sum、avg、max、min等聚合查询加速有20%～30%提速
2. 对sum加速10%~20%
3. 可能是Doris zoneMap的影响，对单查询max、min加速不明显
4. 对count distinct和count(*)没有作用
5. 由于TPCH没有两种特殊类型的列，hll 和 bitmap没测，预计也有提速效果

综合查询图：
<img width="743" alt="restrict test" src="https://user-images.githubusercontent.com/35688959/146346388-63d17c92-7e35-4f15-b975-b2cc8c1bfe51.png">

- 综合查询测试
```
select avg(O_ORDERKEY),sum(O_CUSTKEY),sum(O_TOTALPRICE),max(O_ORDERDATE),min(O_SHIPPRIORITY) from orders;
+-------------------+------------------+---------------------+--------------------+
| avg(`O_ORDERKEY`) | sum(`O_CUSTKEY`) | sum(`O_TOTALPRICE`) | max(`O_ORDERDATE`) |
+-------------------+------------------+---------------------+--------------------+
|        59999991.5 |   44996636714486 |    4533439297407.27 | 1998-08-02         |
+-------------------+------------------+---------------------+--------------------+

不开启向量化引擎: 2.59s 2.75s 2.68s 2.54s 2.67s 平均值 2.65s
优化前: 0.51s 0.59s 0.51s 0.54s 0.61s 平均值 0.55s
优化后: 0.41s 0.34s 0.35s 0.37s 0.40s 平均值 0.37s
```

- 测试sum

```
select sum(O_ORDERKEY),sum(O_TOTALPRICE) from orders;
+-------------------+---------------------+
| sum(`O_ORDERKEY`) | sum(`O_TOTALPRICE`) |
+-------------------+---------------------+
|  1799999745000000 |    4533439297407.27 |
+-------------------+---------------------+

不开启向量化引擎:1.06s 1.06s 1.09s 1.33s 1.26s 平均 1.16s
优化前: 0.21s 0.17s 0.19s 0.23s 0.26s 平均 0.21s
优化后: 0.22s 0.16s 0.19s 0.15s 0.19s 平均 0.18s
```

- 测试count distinct
```
select count(DISTINCT O_ORDERKEY) from orders;
+------------------------------+
| count(DISTINCT `O_ORDERKEY`) |
+------------------------------+
|                     30000000 |
+------------------------------+

不开启向量化引擎: 3.71s 3.89s 3.67s 3.57s 3.66s 平均3.7s
优化前: 2.21s 2.33s 2.32s 2.26s 2.37s 
优化后: 2.21s 2.21s 2.20s 2.25s 2.27s 【数据相近，优化无影响】
```